# api/documentation/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This project is a Node.js api to create your own IMDB. Users are able to add movies, actors and reviews.",
    "version": "1.0.0",
    "title": "Private IMDB API"
  },
...
  "paths": {
...
    "/users/create-user": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Create a user",
        "operationId": "createUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "user data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "$ref": "#/definitions/users"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "insertId": {
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/update-user": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update a user, only the user himself can do this",
        "operationId": "updateUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "user data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "firstname": {
                      "type": "string"
                    },
                    "lastname": {
                      "type": "string"
                    },
                    "profilePicture": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/update-password": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update a user password, only the user himself can do this",
        "operationId": "updateUserPassword",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "user password",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Find a user by id",
        "operationId": "getUserById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "the id of the user to find",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/users"
            }
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/with-details/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Find a user by id with all associations",
        "operationId": "getUserByIdWithAssociations",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "the id of the user to find",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "firstname": {
                  "type": "string"
                },
                "lastname": {
                  "type": "string"
                },
                "profilePicture": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                },
                "movies": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/movies"
                  }
                },
                "reviews": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/users"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, you need a token access to execute this request"
          },
          "412": {
            "description": "Invalid input, missing some parameters in path or body"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "movies": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "rate": {
          "type": "number"
        },
        "date": {
          "type": "string"
        },
        "duration": {
          "type": "integer"
        },
        "genre": {
          "type": "string"
        },
        "synopsis": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "teaserUrl": {
          "type": "string"
        },
        "nationality": {
          "type": "string"
        },
        "userId": {
          "type": "integer"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
    "users": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    }
  }
}